Excluded Thread.destroy() and Thread.stop(Throwable) test cases from JAVA11
Excluded `destroy()` and `stop(Throwable)`  test cases from JAVA11 in
`Thread.java`

[ci skip]
